src/requirements.txt
aiohttp==3.9.3
aiosignal==1.2.0
async-timeout==4.0.3
attrs==23.1.0
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==2.0.4
datasets==2.19.0
dill==0.3.6
filelock==3.13.1
frozenlist==1.4.0
fsspec==2024.3.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.3
numexpr==2.8.7
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pip==23.3.1
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
responses==0.13.3
rsa==4.9
setuptools==68.2.2
six==1.16.0
sympy==1.12
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.65.0
triton==2.3.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wheel==0.41.2
xxhash==2.0.2
yarl==1.9.3
